[Gold III] Title: 웜홀, Time: 348 ms, Memory: 20092 KB -BaekjoonHub

+++ 백준/Gold/1865. 웜홀/웜홀.java
@@ -36,7 +36,7 @@ public class Main {
     }
 
     private static boolean bellmanFord_cycle(List<Edge> graph, int N) {
-        long[] dist = new long[N+1];
+        int[] dist = new int[N+1];
 
         for (int i = 0; i < N; i++) {
             for (Edge e :
